renamed IGraph to IAdjacencyIndex
