+++ src/Graph.Lib/Indexes/DirectedAdjacencyMatrix.cs
@@ -1,9 +1,11 @@
-﻿namespace Graph.Sets
+﻿using Graph.Sets;
+
+namespace Graph.Indexes
 {
     public sealed class DirectedAdjacencyMatrix
         : AdjacencyMatrix
     {
-        public static IGraph Empty { get; } = new DirectedAdjacencyMatrix();
+        public static IAdjacencyIndex Empty { get; } = new DirectedAdjacencyMatrix();
 
         public override GraphType Type => GraphType.Directed;
 
@@ -32,7 +34,7 @@
             this.Matrix[vertex1, vertex2] = 0;
         }
 
-        public override IGraph Resize(int size)
+        public override IAdjacencyIndex Resize(int size)
         {
             return new DirectedAdjacencyMatrix(this, size);
         }
